requirements.txt
requests
selenium
bs4
pyrebase4